Introduce an interface for EventSpyDispatcher
