+++ maven-core/src/main/java/org/apache/maven/eventspy/internal/EventSpyDispatcher.java
@@ -19,131 +19,25 @@ package org.apache.maven.eventspy.internal;
  * under the License.
  */
 
-import java.util.ArrayList;
-import java.util.List;
-
-import javax.inject.Inject;
-import javax.inject.Named;
-import javax.inject.Singleton;
-
 import org.apache.maven.eventspy.EventSpy;
 import org.apache.maven.execution.ExecutionListener;
-import org.codehaus.plexus.logging.Logger;
 import org.eclipse.aether.RepositoryListener;
 
 /**
  * Dispatches callbacks to all registered eventspies.
  * @since 3.0.2
  */
-@Named
-@Singleton
-public class EventSpyDispatcher
+public interface EventSpyDispatcher
 {
 
-    @Inject
-    private Logger logger;
+    ExecutionListener chainListener( ExecutionListener listener );
 
-    private List<EventSpy> eventSpies;
+    RepositoryListener chainListener( RepositoryListener listener );
 
-    @Inject
-    public void setEventSpies( List<EventSpy> eventSpies )
-    {
-        // make copy to get rid of needless overhead for dynamic lookups
-        this.eventSpies = new ArrayList<>( eventSpies );
-    }
+    void init( EventSpy.Context context );
 
-    public List<EventSpy> getEventSpies()
-    {
-        return eventSpies;
-    }
+    void onEvent( Object event );
 
-    public ExecutionListener chainListener( ExecutionListener listener )
-    {
-        if ( eventSpies.isEmpty() )
-        {
-            return listener;
-        }
-        return new EventSpyExecutionListener( this, listener );
-    }
-
-    public RepositoryListener chainListener( RepositoryListener listener )
-    {
-        if ( eventSpies.isEmpty() )
-        {
-            return listener;
-        }
-        return new EventSpyRepositoryListener( this, listener );
-    }
-
-    public void init( EventSpy.Context context )
-    {
-        if ( eventSpies.isEmpty() )
-        {
-            return;
-        }
-        for ( EventSpy eventSpy : eventSpies )
-        {
-            try
-            {
-                eventSpy.init( context );
-            }
-            catch ( Exception | LinkageError e )
-            {
-                logError( "initialize", e, eventSpy );
-            }
-        }
-    }
-
-    public void onEvent( Object event )
-    {
-        if ( eventSpies.isEmpty() )
-        {
-            return;
-        }
-        for ( EventSpy eventSpy : eventSpies )
-        {
-            try
-            {
-                eventSpy.onEvent( event );
-            }
-            catch ( Exception | LinkageError e )
-            {
-                logError( "notify", e, eventSpy );
-            }
-        }
-    }
-
-    public void close()
-    {
-        if ( eventSpies.isEmpty() )
-        {
-            return;
-        }
-        for ( EventSpy eventSpy : eventSpies )
-        {
-            try
-            {
-                eventSpy.close();
-            }
-            catch ( Exception | LinkageError e )
-            {
-                logError( "close", e, eventSpy );
-            }
-        }
-    }
-
-    private void logError( String action, Throwable e, EventSpy spy )
-    {
-        String msg = "Failed to " + action + " spy " + spy.getClass().getName() + ": " + e.getMessage();
-
-        if ( logger.isDebugEnabled() )
-        {
-            logger.warn( msg, e );
-        }
-        else
-        {
-            logger.warn( msg );
-        }
-    }
+    void close();
 
 }
